requirements.txt
affine==2.4.0
aiofiles==24.1.0
altair==5.5.0
aniso8601==10.0.1
annotated-types==0.7.0
anyio==4.9.0
anywidget==0.9.18
appdirs==1.4.4
appnope==0.1.4
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
bokeh==3.7.3
bqplot==0.12.45
branca==0.8.1
build==1.2.2.post1
cachelib==0.13.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
click-plugins==1.1.1
cligj==0.7.2
color-operations==0.2.0
colorcet==3.1.0
colour==0.1.5
comm==0.2.2
commonmark==0.9.1
contourpy==1.3.2
cycler==0.12.1
dash==3.0.4
debugpy==1.8.14
decorator==5.2.1
docstring_parser==0.16
duckdb==1.3.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.2.0
faicons==0.2.2
fastapi==0.115.12
fastjsonschema==2.21.1
ffmpy==0.5.0
filelock==3.18.0
Flask==3.0.3
Flask-Caching==2.3.1
flask-cors==6.0.0
flask-restx==1.3.0
folium==0.19.6
fonttools==4.58.1
fsspec==2025.5.1
fuzzbucket-client==0.13.0
gast==0.4.0
gdown==5.2.0
geojson==3.2.0
geopandas==1.0.1
gitdb==4.0.12
GitPython==3.1.44
gradio==5.31.0
gradio_client==1.10.1
graphviz==0.20.3
great-tables==0.17.0
groovy==0.1.2
h11==0.16.0
h5netcdf==1.6.1
h5py==3.13.0
hf-xet==1.1.2
holoviews==1.20.2
htmltools==0.6.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.2
humanize==4.12.3
hvplot==0.11.3
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
ipydatagrid==1.4.0
ipyevents==2.0.2
ipyfilechooser==0.6.0
ipykernel==6.29.5
ipyleaflet==0.19.2
ipympl==0.9.7
ipython==8.36.0
ipytree==0.2.2
ipyvue==1.11.2
ipyvuetify==1.11.1
ipywidgets==8.1.7
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter-leaflet==0.19.2
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
leafmap==0.46.9
linkify-it-py==2.0.3
localtileserver==0.10.6
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
morecantile==6.2.0
mpmath==1.3.0
narwhals==1.38.0
nbclient==0.10.2
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
numexpr==2.10.2
numpy==2.2.6
openpyxl==3.1.5
opera-utils==0.21.0
orjson==3.10.18
packaging==24.2
pandas==2.2.3
panel==1.7.1
param==2.2.0
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
pillow==11.2.1
pins==0.8.7
pip-tools==7.4.1
platformdirs==4.3.8
plotly==6.1.2
pluggy==1.6.0
polars==1.30.0
pooch==1.8.2
prompt_toolkit==3.0.51
protobuf==6.31.0
psutil==7.0.0
psycopg2-binary==2.9.10
psygnal==0.13.0
ptyprocess==0.7.0
pure_eval==0.2.3
py2vega==0.6.1
pyarrow==20.0.0
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
pydot==4.0.0
pydub==0.25.1
Pygments==2.19.1
PyJWT==2.10.1
pyogrio==0.11.0
pyparsing==3.2.3
pyproj==3.7.1
pyproject_hooks==1.2.0
pyshp==2.3.1
PySocks==1.7.1
pystac==1.13.0
pystac-client==0.8.6
pytest==8.4.1
pytest-asyncio==1.0.0
python-box==7.3.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
pyviz_comms==3.0.4
PyYAML==6.0.2
pyzmq==26.4.0
questionary==2.1.0
rasterio==1.4.3
referencing==0.36.2
requests==2.32.3
retrying==1.3.4
rich==14.0.0
ridgeplot==0.3.2
rio-cogeo==5.4.1
rio-tiler==7.7.3.post1
rioxarray==0.19.0
rpds-py==0.25.1
rsconnect_python==1.25.2
ruff==0.11.11
safehttpx==0.1.6
scikit-learn==1.6.1
scipy==1.15.3
scooby==0.10.1
seaborn==0.13.2
semantic-version==2.10.0
semver==3.0.4
server-thread==0.3.0
shapely==2.1.1
shellingham==1.5.4
shiny==1.4.0
shinywidgets==0.6.2
shtab==1.7.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
stack-data==0.6.3
starlette==0.46.2
statsmodels==0.14.4
streamlit==1.45.1
streamlit-aggrid==1.1.5
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.2
torch==2.7.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
traittypes==0.2.1
typeguard==4.4.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tyro==0.9.22
tzdata==2025.2
uc-micro-py==1.0.3
urllib3==2.4.0
uvicorn==0.34.2
vetiver==0.2.6
watchfiles==1.0.5
wcwidth==0.2.13
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.0.6
whitebox==2.3.6
whiteboxgui==2.3.0
widgetsnbextension==4.0.14
xarray==2025.4.0
xmltodict==0.14.2
xxhash==3.5.0
xyzservices==2025.4.0
zipp==3.22.0
